Criação de Paginas de Login, funcionalidade de logout e pagina de resgistro de usuario

+++ Python/Django/StudyBud/base/views.py
@@ -1,13 +1,57 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
+
 
 # rooms = [
 #     {'id': 1, 'name': 'Vamos aprender python!'},
 #     {'id': 2, 'name': 'Design comigo'},
 #     {'id': 3, 'name': 'Desenvolvedores Frontend!'},
 # ]
+
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'O Usuário não existe.')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Nome de Usuário OU Senha não existem')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+
+    context = {'page': page}
+    return render(request, 'main/login_register.html', context)
 
 
 def home(request):
@@ -29,6 +73,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -41,9 +86,13 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Sua presença não é permitida aqui!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -54,8 +103,13 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Sua presença não é permitida aqui!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
